Practice/django_galleries/requirements.txt
Django
Pillow
django-allauth
django-imagekit
